Renamed stuff

+++ utils/parse.py
@@ -51,9 +51,10 @@ def parse_csv(relative_file_path):
     csv_data = pandas.read_csv(file_path)
 
     class_set = csv_data.as_matrix(columns = csv_data.columns[-1:])
-    class_set = numpy.array(class_set).astype('int')
+    class_set = numpy.array(class_set).astype('float')
 
     feature_set = csv_data.as_matrix(columns = csv_data.columns[1:])
-    feature_set = numpy.array(feature_set).astype('int')
+    feature_set = numpy.array(feature_set).astype('float')
+    # feature_set = numpy.around(feature_set, decimals = 8)
 
     return class_set, feature_set
